avance con implementacion con angular

+++ src/main/java/com/cibertec/model/repository/AlumnoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.cibertec.model.Alumno;
 
 @Repository
-public interface AlumnoRepository extends JpaRepository<Alumno, Integer>{
+public interface AlumnoRepository extends JpaRepository<Alumno, Long>{
 	
 	Alumno findByUsualumAndPassalum(String Usualum,String Passalum);
